move site change logic to listing panel
